+++ example/lib/main.dart
@@ -1,0 +1,17 @@
+import 'package:example/screens/home_screen.dart';
+import 'package:flutter/material.dart';
+
+void main() {
+  runApp(MyApp());
+}
+
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Scaffold(
+        body: HomeScreen(),
+      ),
+    );
+  }
+}