+++ test/bloc/rx_bloc_base_test.dart
@@ -36,8 +36,8 @@ void main() {
         Result.success(null),
       ]);
       bloc.setLoadingStateHandler(stream);
-      expect(bloc.loadingState,
-          emitsInOrder([false, true, false, true, false]));
+      expect(
+          bloc.loadingState, emitsInOrder([false, true, false, true, false]));
     });
 
     test('Exception occured', () async {
@@ -54,22 +54,33 @@ void main() {
     test('Test setResultStateHandler with shareStream enabled', () async {
       int listenCount = 0;
       final bloc = BlocImpl();
-      final stream = Future.value(3.14).asResultStream()
-          .doOnListen(() { listenCount++; });
-      final data = bloc.setResultStateHandler(stream, shareStream: true)
-          .whereSuccess();
+      final stream =
+          Future.delayed(Duration(milliseconds: 300), () => Future.value(3.14))
+              .asResultStream()
+              .doOnListen(() {
+        listenCount++;
+      });
+      final data =
+          bloc.setResultStateHandler(stream, shareStream: true).whereSuccess();
       expect(bloc.loadingState, emitsInOrder([false, true, false]));
       expect(data, emits(3.14));
       expect(listenCount, 1);
     });
 
-    test('Test setResultStateHandler with shareStream enabled on broadcast stream', () async {
+    test(
+        'Test setResultStateHandler with shareStream enabled on broadcast stream',
+        () async {
       int listenCount = 0;
       final bloc = BlocImpl();
-      final stream = Future.value(3.14).asResultStream().asBroadcastStream()
-          .doOnListen(() { listenCount++; });
-      final data = bloc.setResultStateHandler(stream, shareStream: true)
-          .whereSuccess();
+      final stream =
+          Future.delayed(Duration(milliseconds: 300), () => Future.value(3.14))
+              .asResultStream()
+              .asBroadcastStream()
+              .doOnListen(() {
+        listenCount++;
+      });
+      final data =
+          bloc.setResultStateHandler(stream, shareStream: true).whereSuccess();
       expect(bloc.loadingState, emitsInOrder([false, true, false]));
       expect(data, emits(3.14));
       expect(listenCount, 1);
@@ -78,28 +89,37 @@ void main() {
     test('Test setResultStateHandler with shareStream disabled', () async {
       int listenCount = 0;
       final bloc = BlocImpl();
-      final stream = Future.value(3.14).asResultStream()
-          .doOnListen(() { listenCount++; });
-      final data = bloc.setResultStateHandler(stream, shareStream: false)
-          .whereSuccess();
+      final stream = Future.value(3.14).asResultStream().doOnListen(() {
+        listenCount++;
+      });
+      final data =
+          bloc.setResultStateHandler(stream, shareStream: false).whereSuccess();
       expect(bloc.loadingState, emitsInOrder([]));
       expect(data, emitsInOrder([]));
-      expect(listenCount, 3);
+      expect(listenCount, 1);
     });
 
-    test('Test setResultStateHandler with shareStream disabled on broadcast stream', () async {
+    test(
+        'Test setResultStateHandler with shareStream disabled on broadcast stream',
+        () async {
       int listenCount = 0;
       final bloc = BlocImpl();
-      final stream = Future.value(3.14).asResultStream().asBroadcastStream()
-          .doOnListen(() { listenCount++; });
-      final data = bloc.setResultStateHandler(stream, shareStream: false)
-          .whereSuccess();
+      final stream = Future.value(3.14)
+          .asResultStream()
+          .asBroadcastStream()
+          .doOnListen(() {
+        listenCount++;
+      });
+      final data =
+          bloc.setResultStateHandler(stream, shareStream: false).whereSuccess();
       expect(bloc.loadingState, emitsInOrder([]));
       expect(data, emitsInOrder([]));
-      expect(listenCount, 3);
+      expect(listenCount, 1);
     });
 
-    test('Test setResultStateHandler with shareStream enabled doesn\'t share the stream again', () async {
+    test(
+        'Test setResultStateHandler with shareStream enabled doesn\'t share the stream again',
+        () async {
       final bloc = BlocImpl();
       final stream = Future.value(3.14).asResultStream().share();
       final streamAfterSettingHandlers =