+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## [0.3.1] - April 29, 2020
+* Upgraded to last stable Flutter/Dart
+* Upgraded to last stable RxDart
+* Added an example
+* Stability improvements
+
 ## [0.3.0] - April 5, 2020
 
 * Added `asResultStream()` extension method to the Stream
-* Fix the behaviour of `setResultStateHandler` when `shareStream` parameter is `true`.
+* Fixed behaviour of `setResultStateHandler` when `shareStream` parameter is `true`.
 
 ## [0.2.0] - March 18, 2020